Renames geometry_util package to swri_geometry_util. Refs #231.
