+++ swri_geometry_util/package.xml
@@ -1,9 +1,9 @@
 <package format="2">
-  <name>geometry_util</name>
+  <name>swri_geometry_util</name>
   <version>0.0.1</version>
   <description>
 
-     geometry_util
+     swri_geometry_util
 
   </description>
   <author>Marc Alban</author>
